Relocate pfs/drive to pfs/server.
