+++ src/server/cmd/pachd/main.go
@@ -19,7 +19,6 @@ import (
 	ppsclient "github.com/pachyderm/pachyderm/src/client/pps"
 	"github.com/pachyderm/pachyderm/src/client/version"
 	"github.com/pachyderm/pachyderm/src/server/health"
-	pfs_driver "github.com/pachyderm/pachyderm/src/server/pfs/drive"
 	pfs_server "github.com/pachyderm/pachyderm/src/server/pfs/server"
 	cache_pb "github.com/pachyderm/pachyderm/src/server/pkg/cache/groupcachepb"
 	cache_server "github.com/pachyderm/pachyderm/src/server/pkg/cache/server"
@@ -142,7 +141,7 @@ func do(appEnvObj interface{}) error {
 	if err != nil {
 		return err
 	}
-	driver, err := pfs_driver.NewDriver(address, []string{etcdAddress}, appEnv.PFSEtcdPrefix, pfsCacheBytes)
+	driver, err := pfs_server.NewDriver(address, []string{etcdAddress}, appEnv.PFSEtcdPrefix, pfsCacheBytes)
 	if err != nil {
 		return err
 	}
