+++ src/server/pfs/server/server.go
@@ -2,7 +2,6 @@ package server
 
 import (
 	pfsclient "github.com/pachyderm/pachyderm/src/client/pfs"
-	"github.com/pachyderm/pachyderm/src/server/pfs/drive"
 	"github.com/pachyderm/pachyderm/src/server/pkg/metrics"
 	"github.com/pachyderm/pachyderm/src/server/pkg/obj"
 )
@@ -32,7 +31,7 @@ type BlockAPIServer interface {
 }
 
 // NewAPIServer creates an APIServer.
-func NewAPIServer(driver drive.Driver, reporter *metrics.Reporter) APIServer {
+func NewAPIServer(driver *driver, reporter *metrics.Reporter) APIServer {
 	return newAPIServer(driver, reporter)
 }
